requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==21.4.0
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.10.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
colorama==0.4.5
cryptography==36.0.2
decorator==5.1.1
distlib==0.3.4
Django==4.1.5
django-summernote==0.8.20.0
djangorestframework==3.14.0
filelock==3.4.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
jsonpath-ng==1.5.3
outcome==1.1.0
packaging==21.3
platformdirs==2.4.0
ply==3.11
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
pytz==2022.7
requests==2.27.1
rfc3986==1.5.0
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
speedtest-cli==2.1.3
sqlparse==0.4.3
trio==0.20.0
trio-websocket==0.9.2
tzdata==2022.7
urllib3==1.26.9
virtualenv==20.10.0
webencodings==0.5.1
wsproto==1.1.0
